+++ lib/nghttp3_stream.h
@@ -253,7 +253,7 @@ struct nghttp3_stream {
   /* pp is nghttp3_push_promise that this stream fulfills. */
   nghttp3_push_promise *pp;
   /* error_code indicates the reason of closure of this stream. */
-  uint16_t error_code;
+  uint64_t error_code;
 
   struct {
     nghttp3_stream_http_state hstate;
